refactor(pcp): split index.rs into graph + resolve modules
Extract the composition graph types (Node, NodeId, NodeFlags, ArcType,
PrimIndexGraph and the strength-order projection) into a new
pcp/graph.rs, and the PrimIndex value-resolution methods (opinions and
the resolve_* field walks) into a new pcp/resolve.rs. index.rs keeps the
IndexBuilder driver, the PrimIndex struct with its build/relocate entry
points, and CompositionContext.

Fields and methods used across the new module boundary are widened to
pub(crate); re-exports move to pcp::graph and consumer imports in
cache/deps/rel are repointed. Pure code relocation: the public API and
PrimIndex::build_with_cache boundary are unchanged, and all 285
composition goldens plus the full suite stay green.
