+++ src/pcp/index.rs
@@ -4,691 +4,20 @@
 //! contribute opinions to a single composed prim. See the
 //! [module-level docs](super) for the full composition overview.
 
-use std::borrow::Cow;
 use std::cmp::Ordering;
 use std::collections::hash_map::Entry;
 use std::collections::{HashMap, HashSet};
 use std::path::PathBuf;
 
 use anyhow::Result;
-use bitflags::bitflags;
 
 use crate::ar::Resolver;
 use crate::sdf::schema::{ChildrenKey, FieldKey};
-use crate::sdf::{self, AbstractData, LayerOffset, ListOp, Path, Payload, PayloadListOp, Reference, Specifier, Value};
+use crate::sdf::{self, AbstractData, LayerOffset, ListOp, Path, Payload, PayloadListOp, Reference, Value};
 
-use super::clip;
+use super::graph::{ArcType, Node, NodeFlags, NodeId, PrimIndexGraph};
 use super::mapping::MapFunction;
 use super::{Error, LayerStack, VariantFallbackMap};
-
-/// The type of composition arc that introduced a [`Node`].
-///
-/// Variants are ordered by LIVERPS strength (strongest first). The
-/// derived `PartialOrd`/`Ord` use the discriminant, so
-/// `Root < Inherit < … < Specialize`.
-#[repr(u8)]
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub enum ArcType {
-    /// Direct opinions from the root layer stack (sublayers).
-    Root,
-    /// Inherited from a class prim.
-    Inherit,
-    /// Contributed by a selected variant.
-    Variant,
-    /// Contributed by a relocate (non-destructive namespace remapping).
-    Relocate,
-    /// Brought in via a reference arc.
-    Reference,
-    /// Brought in via a payload arc.
-    Payload,
-    /// Contributed by a specializes arc (weakest).
-    Specialize,
-}
-
-/// Stable handle to a [`Node`] within a [`PrimIndex`]'s composition graph
-/// (C++ `PcpNodeRef`).
-///
-/// A handle stays valid for the life of the index: the node arena is only
-/// ever appended to, never reordered, so a `NodeId` is safe to store in
-/// another node's `parent`/`children`/`origin`. The sentinel value
-/// [`INVALID`](Self::INVALID) represents the absence of a node.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct NodeId(u32);
-
-impl NodeId {
-    /// Sentinel: no node.
-    pub const INVALID: Self = Self(u32::MAX);
-
-    /// Returns `true` if this handle points to an actual node.
-    pub fn is_valid(self) -> bool {
-        self.0 != u32::MAX
-    }
-
-    /// Converts to a `usize` for indexing into the arena.
-    pub(crate) fn idx(self) -> usize {
-        self.0 as usize
-    }
-}
-
-impl Default for NodeId {
-    /// The default handle is the [`INVALID`](Self::INVALID) sentinel.
-    fn default() -> Self {
-        Self::INVALID
-    }
-}
-
-bitflags! {
-    /// Per-node composition state, mirroring the booleans carried by C++
-    /// `PcpNodeRef`.
-    ///
-    /// Most bits are reserved for parity features not yet wired up; only the
-    /// flags set during composition today affect resolution. Reserving the
-    /// full surface now keeps later work from re-editing [`Node`].
-    #[derive(Debug, Default, Clone, Copy, PartialEq, Eq)]
-    pub struct NodeFlags: u16 {
-        /// Contributes no opinions; kept only for graph structure.
-        const INERT = 1 << 0;
-        /// Hidden from value resolution but retained for change tracking.
-        const CULLED = 1 << 1;
-        /// Subtree namespace-restricted by a relocate.
-        const RESTRICTED = 1 << 2;
-        /// Blocked by `permission = private` on a stronger site.
-        const PERMISSION_DENIED = 1 << 3;
-        /// This site is itself `permission = private`.
-        const PERMISSION_PRIVATE = 1 << 4;
-        /// Children are prohibited (e.g. an unloaded payload).
-        const PROHIBITED_CHILDREN = 1 << 5;
-        /// Added by implied inherit/specialize propagation.
-        const IMPLIED_CLASS = 1 << 6;
-        /// Introduced by a directly-authored arc (not implied).
-        const DIRECT = 1 << 7;
-        /// Introduced through a specializes arc; globally weak (spec 10.4.1).
-        const HAS_SPECIALIZES = 1 << 8;
-        /// Introduced by a relocate.
-        const RELOCATE_SOURCE = 1 << 9;
-        /// Lies inside a selected variant branch.
-        const VARIANT_BRANCH = 1 << 10;
-    }
-}
-
-/// A single node in the composition graph (C++ `PcpNodeRef`).
-///
-/// Each node represents a site (layer + path) contributing opinions to a
-/// composed prim. The identity fields (`layer_index`, `path`, `arc`)
-/// define *what* this node contributes. The namespace mappings
-/// (`map_to_parent`, `map_to_root`) translate paths across composition arcs.
-/// The structure fields (`parent`, `children`, `origin`) record the node's
-/// place in the composition tree; access them through [`PrimIndex`].
-#[derive(Debug, Clone)]
-pub struct Node {
-    /// The layer stack contributing opinions at this site (C++
-    /// `PcpNode::GetLayerStack`): each member is a `(layer index, sublayer
-    /// offset)` pair, strongest sublayer first. A member's sublayer offset is
-    /// composed atop `map_to_root`'s arc offset during value resolution (see
-    /// [`layers`](Self::layers)).
-    layer_stack: Vec<(usize, LayerOffset)>,
-    /// The path within that layer (may differ from composed path due to remapping).
-    pub path: Path,
-    /// The composition arc that introduced this node.
-    pub arc: ArcType,
-    /// Maps paths from this node's namespace to its parent's namespace.
-    pub map_to_parent: MapFunction,
-    /// Maps paths from this node's namespace directly to the root namespace.
-    /// Computed as `parent.map_to_root.compose(self.map_to_parent)`.
-    pub map_to_root: MapFunction,
-    /// Structural parent in the composition tree, or `None` for a root node.
-    parent: Option<NodeId>,
-    /// Structural children, in the order they were introduced (strength order
-    /// among siblings).
-    children: Vec<NodeId>,
-    /// Node that introduced this arc (C++ `GetOriginNode`): the parent for a
-    /// direct arc (set at [`add_child`](PrimIndexGraph::add_child) time), or
-    /// the propagated-from node for an implied class or graft. `None` only for
-    /// the synthetic root, which has no parent.
-    origin: Option<NodeId>,
-    /// Namespace depth at which the introducing arc was authored. Used by
-    /// implied inherits/specializes that propagate toward the root.
-    namespace_depth: u16,
-    /// Position of this node's prim in its specializes chain: 0 for a directly
-    /// specialized class, 1 for the class it specializes, and so on. A specialize
-    /// source is weaker than its target (spec 10.4.1), so a lower depth is
-    /// stronger. Set during [`finalize_strength_order`](PrimIndexGraph::finalize_strength_order);
-    /// meaningful only for specialize-introduced nodes.
-    specialize_chain_depth: u16,
-    /// Composition state bits (see [`NodeFlags`]).
-    flags: NodeFlags,
-}
-
-impl Node {
-    /// Builds a standalone node with no structural links.
-    ///
-    /// Callers that append through [`PrimIndexGraph::add_child`] have their
-    /// `parent`/`children` populated by the builder; the relocate inserts and
-    /// grafts set the links explicitly afterward.
-    pub(crate) fn new(
-        layer_index: usize,
-        path: Path,
-        arc: ArcType,
-        map_to_parent: MapFunction,
-        map_to_root: MapFunction,
-        introduced_by_specialize: bool,
-    ) -> Self {
-        let flags = if introduced_by_specialize {
-            NodeFlags::HAS_SPECIALIZES
-        } else {
-            NodeFlags::empty()
-        };
-        Self {
-            // A node lists one layer until the per-site model folds a whole
-            // sublayer stack into it; that lone layer's sublayer offset is
-            // already baked into `map_to_root`, so its entry offset is identity.
-            layer_stack: vec![(layer_index, LayerOffset::IDENTITY)],
-            path,
-            arc,
-            map_to_parent,
-            map_to_root,
-            parent: None,
-            children: Vec::new(),
-            origin: None,
-            namespace_depth: 0,
-            specialize_chain_depth: 0,
-            flags,
-        }
-    }
-
-    /// Builds a standalone node spanning a whole sublayer stack: `layer_stack`
-    /// lists every contributing `(layer index, sublayer offset)` member,
-    /// strongest first. Like [`new`](Self::new) but for a per-site node not
-    /// appended through [`PrimIndexGraph::add_site_child`] (the relocate
-    /// inserts). Panics on an empty `layer_stack`.
-    pub(crate) fn new_site(
-        layer_stack: Vec<(usize, LayerOffset)>,
-        path: Path,
-        arc: ArcType,
-        map_to_parent: MapFunction,
-        map_to_root: MapFunction,
-        introduced_by_specialize: bool,
-    ) -> Self {
-        let mut node = Self::new(
-            layer_stack[0].0,
-            path,
-            arc,
-            map_to_parent,
-            map_to_root,
-            introduced_by_specialize,
-        );
-        node.layer_stack = layer_stack;
-        node
-    }
-
-    /// Index of the strongest layer contributing at this site. A representative
-    /// for callers that key on a single layer (dependencies, dumps); value
-    /// resolution iterates [`layers`](Self::layers) instead.
-    pub fn layer_index(&self) -> usize {
-        self.layer_stack[0].0
-    }
-
-    /// The site's contributing layers as stored: `(layer index, sublayer
-    /// offset)` members, strongest first. Used when grafting a node into
-    /// another index, which copies the stack verbatim. Value resolution should
-    /// use [`layers`](Self::layers) instead, which folds in the arc offset.
-    pub(crate) fn layer_stack(&self) -> &[(usize, LayerOffset)] {
-        &self.layer_stack
-    }
-
-    /// Iterates the site's contributing layers strongest first, each paired
-    /// with its effective time offset to the root namespace: `map_to_root`'s
-    /// arc offset with the member's sublayer offset composed on top. Value
-    /// resolution reads opinions through this iterator so one per-site node can
-    /// fold every sublayer.
-    pub fn layers(&self) -> impl Iterator<Item = (usize, LayerOffset)> + '_ {
-        let arc_offset = self.map_to_root.time_offset();
-        self.layer_stack
-            .iter()
-            .map(move |&(li, sub)| (li, arc_offset.concatenate(&sub)))
-    }
-
-    /// Structural parent, or `None` for a root node.
-    pub fn parent(&self) -> Option<NodeId> {
-        self.parent
-    }
-
-    /// Structural children, in strength order among siblings.
-    pub fn children(&self) -> &[NodeId] {
-        &self.children
-    }
-
-    /// Node that introduced this arc (C++ `GetOriginNode`): the parent for a
-    /// direct arc, the propagated-from node for an implied class or graft, or
-    /// `None` for the synthetic root.
-    pub fn origin(&self) -> Option<NodeId> {
-        self.origin
-    }
-
-    /// Namespace depth at which the introducing arc was authored.
-    pub fn namespace_depth(&self) -> u16 {
-        self.namespace_depth
-    }
-
-    /// Composition state bits.
-    pub fn flags(&self) -> NodeFlags {
-        self.flags
-    }
-
-    /// True when this node contributes no opinions and exists only to give the
-    /// composition graph structure (the synthetic tree root).
-    pub fn is_inert(&self) -> bool {
-        self.flags.contains(NodeFlags::INERT)
-    }
-
-    /// True when this node is retained for composition structure (so an editor
-    /// or change tracking can see its arc) but contributes no opinions to value
-    /// resolution. Set on an arc whose target site authors no spec (C++
-    /// `PcpPrimIndex` culling).
-    pub fn is_culled(&self) -> bool {
-        self.flags.contains(NodeFlags::CULLED)
-    }
-
-    /// True when this node was introduced through a specializes arc (directly
-    /// or transitively), making it globally weak per spec section 10.4.1.
-    pub(crate) fn introduced_by_specialize(&self) -> bool {
-        self.flags.contains(NodeFlags::HAS_SPECIALIZES)
-    }
-
-    /// True when this node is a direct arc to a `permission = private` site, or
-    /// lies in such an arc's subtree (spec 10.3.3). It stays visible
-    /// structurally (`nodes`, `has_spec`, child names) but contributes no
-    /// opinions to value resolution — the C++ `_InertSubtree` behavior.
-    pub(crate) fn is_permission_denied(&self) -> bool {
-        self.flags.contains(NodeFlags::PERMISSION_DENIED)
-    }
-}
-
-/// Arena-based composition graph.
-///
-/// `nodes` is the node arena in insertion (structural) order: it is only ever
-/// appended to, never reordered, so a [`NodeId`] (an index into it) stays
-/// valid for the life of the index and is safe to store in another node's
-/// `parent`/`children`/`origin`. Strongest-to-weakest strength order is a
-/// separate projection, `strength_order`, holding the same handles permuted —
-/// value resolution walks it, not the arena. Dereferencing the graph yields
-/// the arena slice for the builder's structural lookups.
-///
-/// `root` is the synthetic, inert tree root created by
-/// [`init_root`](Self::init_root): every otherwise-parentless node attaches
-/// under it, so the graph is a single rooted tree rather than a forest. It is
-/// [`NodeId::INVALID`] for the hand-built test graphs that never call
-/// `init_root`.
-#[derive(Debug, Clone, Default)]
-pub(crate) struct PrimIndexGraph {
-    nodes: Vec<Node>,
-    strength_order: Vec<NodeId>,
-    root: NodeId,
-}
-
-impl PrimIndexGraph {
-    fn len(&self) -> usize {
-        self.nodes.len()
-    }
-
-    /// Returns the prim's local root node — the `Root`-arc child of the
-    /// synthetic inert root, carrying the prim's own (sublayer) opinions. An
-    /// implied class anchors here so it ranks among the prim's direct arcs (C++
-    /// `_AddClassBasedArc` adds it under the node owning the prim), regardless
-    /// of how deep the arc that implied it sits or whether an ancestral arc was
-    /// grafted as a separate root-level sibling. [`NodeId::INVALID`] when the
-    /// prim has no local opinion (composed purely through arcs).
-    fn local_root(&self) -> NodeId {
-        if !self.root.is_valid() {
-            return NodeId::INVALID;
-        }
-        self.nodes[self.root.idx()]
-            .children
-            .iter()
-            .copied()
-            .find(|&c| self.nodes[c.idx()].arc == ArcType::Root)
-            .unwrap_or(NodeId::INVALID)
-    }
-
-    /// Creates the synthetic, inert tree root (C++ unified graph root) and
-    /// records it as [`root`](Self::root). Must be the first node, so it takes
-    /// [`NodeId`] 0.
-    ///
-    /// Its identity `map_to_parent`/`map_to_root` make re-parenting an
-    /// otherwise-parentless node under it offset-neutral
-    /// (`identity ∘ child.map_to_parent == child.map_to_parent`), so a former
-    /// forest root keeps the `map_to_root` it had with no parent. The node is
-    /// flagged [`INERT`](NodeFlags::INERT) and skipped by value resolution.
-    fn init_root(&mut self, path: Path) -> NodeId {
-        debug_assert!(self.nodes.is_empty(), "synthetic root must be the first node");
-        let id = NodeId(self.nodes.len() as u32);
-        let depth = path.prim_element_count() as u16;
-        let mut node = Node::new(
-            0,
-            path,
-            ArcType::Root,
-            MapFunction::identity(),
-            MapFunction::identity(),
-            false,
-        );
-        node.namespace_depth = depth;
-        node.flags |= NodeFlags::INERT;
-        self.nodes.push(node);
-        self.root = id;
-        id
-    }
-
-    /// Appends a node to the graph, linking it under `parent`.
-    ///
-    /// `parent` identifies the structural parent in the composition graph. An
-    /// [`INVALID`](NodeId::INVALID) `parent` attaches the node under the
-    /// synthetic [`root`](Self::root) when one exists, keeping the graph a
-    /// single tree; `map_to_root` then composes through that identity-mapped
-    /// root, leaving it equal to `map_to_parent`. The new node is recorded in
-    /// its parent's children and appended to the arena; the strength projection
-    /// is built once at the end of the build by
-    /// [`finalize_strength_order`](Self::finalize_strength_order). The returned
-    /// handle stays valid for the life of the index.
-    fn add_child(
-        &mut self,
-        parent: NodeId,
-        layer_index: usize,
-        path: Path,
-        arc: ArcType,
-        map_to_parent: MapFunction,
-        introduced_by_specialize: bool,
-    ) -> NodeId {
-        // A caller-supplied `INVALID` parent means "a root site": the arc is
-        // introduced at this prim. Such nodes re-parent under the synthetic
-        // root for structure, but their namespace depth still derives from
-        // their own path (the conceptual parent site), not the synthetic root.
-        let root_site = !parent.is_valid();
-        let struct_parent = if root_site { self.root } else { parent };
-
-        let map_to_root = if struct_parent.is_valid() {
-            self.nodes[struct_parent.idx()].map_to_root.compose(&map_to_parent)
-        } else {
-            map_to_parent.clone()
-        };
-        // Namespace depth is the prim-element count of the parent site path at
-        // arc introduction (C++ `PcpNode_GetNonVariantPathElementCount`); a
-        // root site uses its own path.
-        let namespace_depth = if root_site {
-            path.prim_element_count()
-        } else {
-            self.nodes[parent.idx()].path.prim_element_count()
-        } as u16;
-        let idx = NodeId(self.nodes.len() as u32);
-        let mut node = Node::new(
-            layer_index,
-            path,
-            arc,
-            map_to_parent,
-            map_to_root,
-            introduced_by_specialize,
-        );
-        node.namespace_depth = namespace_depth;
-        if struct_parent.is_valid() {
-            node.parent = Some(struct_parent);
-            // A directly-authored arc's origin is its parent (C++
-            // `GetOriginNode`). Implied-class and graft propagation overwrite
-            // this afterward; setting it here makes `origin` total.
-            node.origin = Some(struct_parent);
-            self.nodes[struct_parent.idx()].children.push(idx);
-        }
-        self.nodes.push(node);
-        idx
-    }
-
-    /// Like [`add_child`](Self::add_child) but for a site spanning several
-    /// sublayers: `layer_stack` lists every contributing `(layer index,
-    /// sublayer offset)`, strongest sublayer first. The first member is the
-    /// site representative; the remaining members are folded into the node's
-    /// layer stack so value resolution reads each sublayer in turn. Panics on
-    /// an empty `layer_stack`.
-    fn add_site_child(
-        &mut self,
-        parent: NodeId,
-        layer_stack: Vec<(usize, LayerOffset)>,
-        path: Path,
-        arc: ArcType,
-        map_to_parent: MapFunction,
-        introduced_by_specialize: bool,
-    ) -> NodeId {
-        let id = self.add_child(
-            parent,
-            layer_stack[0].0,
-            path,
-            arc,
-            map_to_parent,
-            introduced_by_specialize,
-        );
-        self.nodes[id.idx()].layer_stack = layer_stack;
-        id
-    }
-
-    /// Tags a node as introduced by implied-class propagation, recording the
-    /// originating arc node. Pure metadata: it does not affect resolution.
-    fn mark_implied(&mut self, id: NodeId, origin: NodeId) {
-        let node = &mut self.nodes[id.idx()];
-        node.flags |= NodeFlags::IMPLIED_CLASS;
-        if origin.is_valid() {
-            node.origin = Some(origin);
-        }
-    }
-
-    /// Collects the chain of nodes from `id` up to its tree root, node first
-    /// and root last.
-    fn chain_to_root(&self, id: NodeId) -> Vec<NodeId> {
-        let mut chain = vec![id];
-        let mut cur = id;
-        while let Some(parent) = self.nodes[cur.idx()].parent {
-            chain.push(parent);
-            cur = parent;
-        }
-        chain
-    }
-
-    /// Compares two sibling nodes (nodes sharing a parent) by strength,
-    /// porting C++ `PcpCompareSiblingNodeStrength`. Returns [`Ordering::Less`]
-    /// when `a` is the stronger sibling.
-    ///
-    /// Keys, in priority order (spec 10.3): arc type (lower discriminant
-    /// stronger); for two specializes, specializes-chain depth (a source is
-    /// weaker than its target); namespace depth (deeper stronger, skipped for
-    /// specializes); origin strength (the node whose origin is reached first in
-    /// a strong-to-weak walk); finally sibling number — the earlier-created
-    /// node (lower [`NodeId`]) is stronger.
-    fn compare_sibling_node_strength(&self, a: NodeId, b: NodeId) -> Ordering {
-        let na = &self.nodes[a.idx()];
-        let nb = &self.nodes[b.idx()];
-
-        // 1. Arc type: lower discriminant (e.g. Inherit) is stronger.
-        if na.arc != nb.arc {
-            return na.arc.cmp(&nb.arc);
-        }
-
-        let both_specialize = na.introduced_by_specialize() && nb.introduced_by_specialize();
-
-        // 2a. Specializes-chain depth: a class is stronger than the class it
-        // specializes (C++ `_OriginsAreNestedArcs`: the source of a specializes
-        // arc is weaker than its target regardless of namespace depth, spec
-        // 10.4.1). Lower depth is stronger.
-        if both_specialize && na.specialize_chain_depth != nb.specialize_chain_depth {
-            return na.specialize_chain_depth.cmp(&nb.specialize_chain_depth);
-        }
-
-        // 2b. Namespace depth: a deeper arc-introduction site is stronger. C++
-        // skips this for specializes whose origins are nested arcs (handled by
-        // the chain-depth key above), so it is skipped for two specializes.
-        if !both_specialize && na.namespace_depth != nb.namespace_depth {
-            return nb.namespace_depth.cmp(&na.namespace_depth);
-        }
-
-        // 3. Origin strength, only when the two origins differ. `origin` is
-        // total (C++ `GetOriginNode`): the node an implied arc was propagated
-        // from, or the parent for a directly-authored arc, so two direct
-        // siblings share an origin and fall through to the tiebreak below. The
-        // synthetic root is the sole node without an origin; it stands in for
-        // itself. Comparing origins recurses only over strictly-older nodes (an
-        // origin is always created before the node carrying it), so the
-        // recursion is well-founded — except when both nodes default to
-        // themselves (no origin authored): recursing on the same `(a, b)` would
-        // not progress, so skip straight to the sibling-number tiebreak.
-        let oa = na.origin.unwrap_or(a);
-        let ob = nb.origin.unwrap_or(b);
-        if oa != ob && (oa != a || ob != b) {
-            let ord = self.compare_node_strength(oa, ob);
-            if ord != Ordering::Equal {
-                return ord;
-            }
-        }
-
-        // 4. Sibling number: the earlier-created node (lower arena index) was
-        // introduced first and is stronger.
-        a.0.cmp(&b.0)
-    }
-
-    /// Compares two arbitrary nodes in the same tree by strength, porting C++
-    /// `PcpCompareNodeStrength`. Returns [`Ordering::Less`] when `a` is stronger.
-    ///
-    /// Walks each node's chain to the shared root to find their lowest common
-    /// ancestor. When one node is an ancestor of the other, the ancestor is
-    /// stronger; otherwise the two diverging children are siblings, compared by
-    /// [`compare_sibling_node_strength`](Self::compare_sibling_node_strength).
-    fn compare_node_strength(&self, a: NodeId, b: NodeId) -> Ordering {
-        if a == b {
-            return Ordering::Equal;
-        }
-        let chain_a = self.chain_to_root(a);
-        let chain_b = self.chain_to_root(b);
-
-        // Walk both chains from the root downward (chains are leaf-first) until
-        // they diverge. The last shared node is the lowest common ancestor.
-        let mut ia = chain_a.len();
-        let mut ib = chain_b.len();
-        while ia > 0 && ib > 0 {
-            let ca = chain_a[ia - 1];
-            let cb = chain_b[ib - 1];
-            if ca != cb {
-                // First divergence: `ca` and `cb` are siblings under the LCA.
-                return self.compare_sibling_node_strength(ca, cb);
-            }
-            ia -= 1;
-            ib -= 1;
-        }
-
-        // One chain is a prefix of the other: the shorter (the ancestor) wins.
-        match (ia, ib) {
-            (0, 0) => Ordering::Equal,
-            (0, _) => Ordering::Less,
-            (_, 0) => Ordering::Greater,
-            _ => Ordering::Equal,
-        }
-    }
-
-    /// Rebuilds the strength-order projection: a pre-order DFS of the tree for
-    /// non-specialize nodes, followed by the globally-weak specialize band
-    /// (spec 10.4.1). The result runs strongest to weakest; reversing it gives
-    /// weak-to-strong.
-    ///
-    /// Each node's `children` are first sorted by
-    /// [`compare_sibling_node_strength`](Self::compare_sibling_node_strength),
-    /// so the DFS yields strength order among siblings. The specialize band is
-    /// then ordered by the same comparator over the whole band: in C++ the
-    /// specializes nodes are copied to the root and ordered there as siblings
-    /// (`_EvalImpliedSpecializes`); ordering them directly with the sibling
-    /// comparator reproduces that without mutating the tree, so the nested
-    /// structure the grafts depend on is preserved.
-    ///
-    /// The comparator reads only arc type, specializes-chain depth, namespace
-    /// depth, origin chains, and arena order — never the projection being
-    /// computed — so it is well-defined here.
-    fn finalize_strength_order(&mut self) {
-        if !self.root.is_valid() {
-            return;
-        }
-
-        self.compute_specialize_chain_depths();
-
-        // Order every node's children by sibling strength. `children` is taken
-        // out so the comparator can borrow the arena immutably during the sort.
-        for i in 0..self.nodes.len() {
-            let mut children = std::mem::take(&mut self.nodes[i].children);
-            children.sort_by(|&a, &b| self.compare_sibling_node_strength(a, b));
-            self.nodes[i].children = children;
-        }
-
-        // Pre-order DFS from the root: visit a node, then its children in
-        // strength order. Pushing children reversed makes the strongest pop
-        // first.
-        let mut dfs = Vec::with_capacity(self.nodes.len());
-        let mut stack = vec![self.root];
-        while let Some(id) = stack.pop() {
-            dfs.push(id);
-            for &child in self.nodes[id.idx()].children.iter().rev() {
-                stack.push(child);
-            }
-        }
-
-        // Specializes are globally weaker than every other opinion (spec
-        // 10.4.1). Keep the non-specialize nodes in DFS order, then append the
-        // specialize band ordered among itself by the sibling comparator (which
-        // treats the band as siblings at the root, mirroring copy-to-root).
-        let (mut order, mut specials): (Vec<NodeId>, Vec<NodeId>) = dfs
-            .into_iter()
-            .partition(|id| !self.nodes[id.idx()].introduced_by_specialize());
-        specials.sort_by(|&a, &b| self.compare_sibling_node_strength(a, b));
-        order.append(&mut specials);
-        self.strength_order = order;
-    }
-
-    /// Records, on every specialize-introduced node, the depth of its prim in
-    /// the specializes chain: the number of specialize ancestors of the deepest
-    /// node at that prim's path. A directly specialized class is depth 0, the
-    /// class it specializes is depth 1, and so on, so a lower depth sorts
-    /// stronger (a specializes source is weaker than its target, spec 10.4.1).
-    ///
-    /// The per-path maximum recovers the chain position even for the implied
-    /// copies a class brings in across references, which hang flat rather than
-    /// nested under the class they specialize.
-    fn compute_specialize_chain_depths(&mut self) {
-        let mut depth_by_path: HashMap<Path, u16> = HashMap::new();
-        for node in &self.nodes {
-            if !node.introduced_by_specialize() {
-                continue;
-            }
-            let mut depth = 0u16;
-            let mut parent = node.parent;
-            while let Some(pid) = parent {
-                let pnode = &self.nodes[pid.idx()];
-                if pnode.introduced_by_specialize() {
-                    depth += 1;
-                }
-                parent = pnode.parent;
-            }
-            depth_by_path
-                .entry(node.path.clone())
-                .and_modify(|d| *d = (*d).max(depth))
-                .or_insert(depth);
-        }
-        for node in &mut self.nodes {
-            if node.introduced_by_specialize() {
-                node.specialize_chain_depth = depth_by_path.get(&node.path).copied().unwrap_or(0);
-            }
-        }
-    }
-}
-
-impl std::ops::Deref for PrimIndexGraph {
-    /// Dereferences to the node arena in insertion (structural) order. The
-    /// builder relies on this for positional lookups while composing; strength
-    /// order is reached through [`PrimIndex::nodes`].
-    type Target = [Node];
-    fn deref(&self) -> &[Node] {
-        &self.nodes
-    }
-}
 
 /// Composition index for a single prim.
 ///
@@ -700,27 +29,6 @@ impl std::ops::Deref for PrimIndexGraph {
 pub struct PrimIndex {
     /// Composition graph: node arena plus strength-order projection.
     graph: PrimIndexGraph,
-}
-
-/// A single authored opinion surfaced by [`PrimIndex::opinions`].
-///
-/// One opinion is yielded per contributing layer of a node's layer stack, so a
-/// per-site node fans out into one opinion per sublayer that authored the
-/// field.
-struct Opinion<'a> {
-    /// The contributing node, strongest-to-weakest in the walk.
-    node: &'a Node,
-    /// Index of the contributing layer within the node's layer stack.
-    layer: usize,
-    /// The path queried in the contributing layer (the node path with the
-    /// property suffix applied).
-    query_path: Cow<'a, Path>,
-    /// The authored value at `query_path`.
-    value: Cow<'a, Value>,
-    /// Effective time offset of the contributing layer to the root namespace
-    /// (the node's arc offset with the layer's sublayer offset composed on
-    /// top). Used to retime time samples and clip schedules.
-    offset: LayerOffset,
 }
 
 impl PrimIndex {
@@ -1146,504 +454,6 @@ impl PrimIndex {
         let mut out: Vec<(String, String)> = selections.into_iter().collect();
         out.sort();
         out
-    }
-
-    /// Resolves a field across the composition graph.
-    ///
-    /// Most fields use strongest-opinion-wins (spec 12.2). Four field classes
-    /// have special rules:
-    ///
-    /// - `specifier`: precedence by `def`/`class`/`over` with direct-inherit handling
-    /// - `variability`: weakest authored opinion wins
-    /// - `custom`: any-true (logical OR across all authored opinions)
-    /// - dictionaries: recursive merge of stronger and weaker dictionary opinions
-    ///
-    /// When `prop_suffix` is `None`, queries use the node's path directly (zero-copy).
-    /// When `Some`, appends the suffix to form a property path for each node.
-    /// A [`Value::ValueBlock`] blocks opinions from weaker layers.
-    pub(crate) fn resolve_field(
-        &self,
-        field: &str,
-        stack: &LayerStack,
-        prop_suffix: Option<&str>,
-    ) -> Result<Option<Value>> {
-        if field == FieldKey::Specifier.as_str() {
-            return self.resolve_specifier(stack, prop_suffix);
-        }
-        if field == FieldKey::Variability.as_str() {
-            return self.resolve_variability(stack, prop_suffix);
-        }
-        if field == FieldKey::Custom.as_str() {
-            return self.resolve_custom(stack, prop_suffix);
-        }
-        if field == FieldKey::TimeSamples.as_str() {
-            return Ok(self.resolve_time_samples(stack, prop_suffix)?.map(Value::TimeSamples));
-        }
-        self.resolve_strongest(field, stack, prop_suffix)
-    }
-
-    /// Resolves a token-list-op field by composing list edits from strongest
-    /// to weakest across all contributing nodes.
-    ///
-    /// This is used for metadata like `apiSchemas`, where the field value is a
-    /// list operation rather than a strongest-opinion scalar. A value block
-    /// stops weaker opinions while preserving any stronger composed edits.
-    pub(crate) fn resolve_token_list_op(
-        &self,
-        field: FieldKey,
-        stack: &LayerStack,
-        prop_suffix: Option<&str>,
-    ) -> Result<Vec<String>> {
-        let field = field.as_str();
-        let mut ops = Vec::new();
-
-        for opinion in self.opinions(field, stack, prop_suffix) {
-            let value = opinion?.value;
-            // TODO: a non-conformant backend may store `apiSchemas` as a plain
-            // `Value::TokenVec` even though the field is declared as a
-            // list-op. We currently skip those opinions (falling through to
-            // `_ => continue`). Tightening this further requires a
-            // schema-aware decode step in the USDC reader (and any other
-            // backend) so list-op fields are always produced as
-            // `Value::TokenListOp`; until that lands, ill-typed opinions for
-            // list-op fields are simply ignored rather than coerced.
-            let list_op = match value.into_owned() {
-                Value::ValueBlock => break,
-                Value::TokenListOp(op) => op,
-                _ => continue,
-            };
-            ops.push(list_op);
-        }
-
-        Ok(compose_list_ops(&ops))
-    }
-
-    /// Resolves a path-list-op field by composing list edits from strongest
-    /// to weakest across all contributing nodes. Used for property-level
-    /// fields like `connectionPaths` and `targetPaths`. A value block stops
-    /// weaker opinions while preserving any stronger composed edits.
-    pub(crate) fn resolve_path_list_op(
-        &self,
-        field: FieldKey,
-        stack: &LayerStack,
-        prop_suffix: Option<&str>,
-    ) -> Result<Vec<Path>> {
-        let field = field.as_str();
-        let mut ops = Vec::new();
-
-        for opinion in self.opinions(field, stack, prop_suffix) {
-            let Opinion {
-                node,
-                query_path,
-                value,
-                ..
-            } = opinion?;
-            // A bare `PathVec` (no list-op envelope) is treated as an explicit
-            // replacement of weaker opinions — the natural interpretation for
-            // a non-list-op-typed value when the field is declared list-op.
-            let list_op = match value.into_owned() {
-                Value::ValueBlock => break,
-                Value::PathListOp(op) => op,
-                Value::PathVec(paths) => sdf::PathListOp::explicit(paths),
-                _ => continue,
-            };
-            ops.push(Self::map_path_list_op_to_root(list_op, &query_path, &node.map_to_root));
-        }
-
-        Ok(compose_list_ops(&ops))
-    }
-
-    /// Translate a path-list-op opinion from one contributing node into the
-    /// composed stage namespace before list-op composition.
-    ///
-    /// Every bucket must be translated, not just contributed values: delete
-    /// and reorder opinions only work when they compare against weaker items
-    /// in the same namespace. Unmappable paths are dropped, matching a
-    /// namespace map whose source domain does not include the authored target.
-    fn map_path_list_op_to_root(op: sdf::PathListOp, anchor: &Path, map: &MapFunction) -> sdf::PathListOp {
-        fn map_paths(paths: Vec<Path>, anchor: &Path, map: &MapFunction) -> Vec<Path> {
-            paths
-                .into_iter()
-                .filter_map(|path| {
-                    // List-op targets are authored in the contributing node's
-                    // namespace; compose them only after translating to the
-                    // stage root namespace so deletes and reorders compare
-                    // like-for-like across layers and arcs.
-                    let absolute = anchor.make_absolute(&path);
-                    map.map_source_to_target(&absolute)
-                })
-                .collect()
-        }
-
-        sdf::PathListOp {
-            explicit: op.explicit,
-            explicit_items: map_paths(op.explicit_items, anchor, map),
-            added_items: map_paths(op.added_items, anchor, map),
-            prepended_items: map_paths(op.prepended_items, anchor, map),
-            appended_items: map_paths(op.appended_items, anchor, map),
-            deleted_items: map_paths(op.deleted_items, anchor, map),
-            ordered_items: map_paths(op.ordered_items, anchor, map),
-        }
-    }
-
-    /// Builds the query path for a node, applying `prop_suffix` if given.
-    /// Borrows the node's path when no suffix is needed (zero-copy).
-    fn query_path<'a>(node: &'a Node, prop_suffix: Option<&str>) -> Result<Cow<'a, Path>> {
-        match prop_suffix {
-            Some(suffix) => Ok(Cow::Owned(Path::new(&format!("{}{suffix}", node.path))?)),
-            None => Ok(Cow::Borrowed(&node.path)),
-        }
-    }
-
-    /// Iterates the authored opinions for `field` across the composition graph,
-    /// strongest to weakest, skipping nodes with no opinion. Centralizes the
-    /// query-path / layer / `try_get` walk shared by every `resolve_*` field
-    /// resolver.
-    fn opinions<'a>(
-        &'a self,
-        field: &'a str,
-        stack: &'a LayerStack,
-        prop_suffix: Option<&'a str>,
-    ) -> impl Iterator<Item = Result<Opinion<'a>>> + 'a {
-        // Each node fans out into one opinion per contributing layer in its
-        // layer stack, strongest sublayer first, so a per-site node surfaces
-        // every sublayer that authored the field. Permission-denied nodes (a
-        // direct arc to a private site, spec 10.3.3) stay in `nodes` for
-        // structural queries but contribute no opinions here.
-        //
-        // TODO(perf): collecting per node allocates a small Vec; a custom
-        // iterator over (node, layer) pairs would avoid it on the hot path.
-        self.nodes()
-            .filter(|node| !node.is_permission_denied())
-            .flat_map(move |node| {
-                let query_path = match Self::query_path(node, prop_suffix) {
-                    Ok(path) => path,
-                    Err(err) => return vec![Err(err)],
-                };
-                let mut out: Vec<Result<Opinion<'a>>> = Vec::new();
-                for (layer, offset) in node.layers() {
-                    match stack.layer(layer).try_get(&query_path, field) {
-                        Ok(Some(value)) => out.push(Ok(Opinion {
-                            node,
-                            layer,
-                            query_path: query_path.clone(),
-                            value,
-                            offset,
-                        })),
-                        Ok(None) => {}
-                        Err(err) => out.push(Err(err)),
-                    }
-                }
-                out
-            })
-    }
-
-    /// Walks nodes from strongest to weakest, returning the first opinion.
-    /// A [`Value::ValueBlock`] returns `None`, blocking weaker layers. When
-    /// the strongest opinion is a dictionary, weaker dictionary opinions are
-    /// recursively merged into it (spec 12.2.5); a `ValueBlock` then blocks
-    /// only the remaining weaker opinions, and weaker non-dictionary opinions
-    /// are ignored.
-    fn resolve_strongest(&self, field: &str, stack: &LayerStack, prop_suffix: Option<&str>) -> Result<Option<Value>> {
-        let mut merged: Option<HashMap<String, Value>> = None;
-        for opinion in self.opinions(field, stack, prop_suffix) {
-            let value = opinion?.value;
-            match (merged.as_mut(), value.into_owned()) {
-                (None, Value::ValueBlock) => return Ok(None),
-                (None, Value::Dictionary(dict)) => merged = Some(dict),
-                (None, other) => return Ok(Some(other)),
-                (Some(_), Value::ValueBlock) => break,
-                (Some(strong), Value::Dictionary(weaker)) => dictionary_over(strong, weaker),
-                (Some(_), _) => {}
-            }
-        }
-        Ok(merged.map(Value::Dictionary))
-    }
-
-    /// Resolves `timeSamples` across the composition graph, applying each
-    /// node's effective layer offset (spec 12.3.2.1) so authored layer time is
-    /// mapped to stage time.
-    ///
-    /// Walks nodes strongest-to-weakest and returns the first node that authors
-    /// time samples, retimed by that node's `map_to_root` offset. A
-    /// [`Value::ValueBlock`] blocks weaker layers, matching [`Self::resolve_strongest`].
-    ///
-    /// Unlike generic fields, time samples never merge across nodes: the
-    /// strongest authored opinion wins as a whole.
-    pub(crate) fn resolve_time_samples(
-        &self,
-        stack: &LayerStack,
-        prop_suffix: Option<&str>,
-    ) -> Result<Option<sdf::TimeSampleMap>> {
-        self.time_samples_in(stack, prop_suffix, None)
-    }
-
-    /// Resolves `timeSamples` considering only opinions from the root layer
-    /// stack (`local_layers`). Used by value-clip resolution, where clip data
-    /// is weaker than the anchoring layer's local opinions but stronger than
-    /// data introduced across reference/payload arcs (spec 12.3.4.5).
-    ///
-    /// Membership is by layer index, not arc type: a referenced layer stack
-    /// also contributes `Root`-arc nodes, so only the stage's own root layer
-    /// stack counts as local.
-    pub(crate) fn resolve_local_time_samples(
-        &self,
-        stack: &LayerStack,
-        prop_suffix: Option<&str>,
-        local_layers: &HashSet<usize>,
-    ) -> Result<Option<sdf::TimeSampleMap>> {
-        self.time_samples_in(stack, prop_suffix, Some(local_layers))
-    }
-
-    /// Shared `timeSamples` walk. When `local_layers` is `Some`, opinions
-    /// whose contributing layer is outside that set are skipped so only
-    /// root-layer-stack opinions contribute.
-    fn time_samples_in(
-        &self,
-        stack: &LayerStack,
-        prop_suffix: Option<&str>,
-        local_layers: Option<&HashSet<usize>>,
-    ) -> Result<Option<sdf::TimeSampleMap>> {
-        let field = FieldKey::TimeSamples.as_str();
-        for opinion in self.opinions(field, stack, prop_suffix) {
-            let Opinion {
-                layer, value, offset, ..
-            } = opinion?;
-            if local_layers.is_some_and(|local| !local.contains(&layer)) {
-                continue;
-            }
-            match value.into_owned() {
-                Value::ValueBlock => return Ok(None),
-                Value::TimeSamples(samples) => {
-                    return Ok(Some(retime_samples(samples, offset)));
-                }
-                _ => continue,
-            }
-        }
-        Ok(None)
-    }
-
-    /// Resolves the `clipSets` strength order, if authored. Returns the ordered
-    /// clip set names (strongest first), folding the list-op edits across every
-    /// contributing layer per generic list-op resolution (spec 12.2.6).
-    ///
-    /// The list op is composed over an empty base, matching C++
-    /// `Usd_ClipSetDefinition`: an authored `clipSets` is the authoritative
-    /// ordered list, so a set absent from it is excluded. This makes the
-    /// return value three-way:
-    ///
-    /// - `None` — no opinion authored anywhere; clip sets fall back to name
-    ///   order (spec 12.3.4.1).
-    /// - `Some([])` — authored but composing to empty (explicit `[]` or a
-    ///   delete that cancels every name); no clip sets are active.
-    /// - `Some(names)` — the composed strength order.
-    ///
-    /// `clipSets` is a string list op; both the `String` and `Token` list-op
-    /// encodings (and bare vecs, treated as explicit) are accepted, since USDC
-    /// backends may decode it either way. A `ValueBlock` with no stronger
-    /// opinion leaves the field unauthored (`None`), falling back to name order.
-    pub(crate) fn clip_sets_order(&self, stack: &LayerStack) -> Result<Option<Vec<String>>> {
-        let mut ops = Vec::new();
-        for opinion in self.opinions(FieldKey::ClipSets.as_str(), stack, None) {
-            match opinion?.value.into_owned() {
-                // Stop weaker opinions while keeping any stronger composed edits.
-                Value::ValueBlock => break,
-                Value::StringListOp(op) | Value::TokenListOp(op) => ops.push(op),
-                Value::StringVec(names) | Value::TokenVec(names) => ops.push(sdf::StringListOp::explicit(names)),
-                _ => continue,
-            }
-        }
-        if ops.is_empty() {
-            return Ok(None);
-        }
-        Ok(Some(compose_list_ops(&ops)))
-    }
-
-    /// Resolves explicit value clip sets while preserving the layer that
-    /// authored path-bearing fields. The top-level `clips` dictionary composes
-    /// recursively, but relative clip assets must still be anchored to the
-    /// layer that supplied `assetPaths`/`manifestAssetPath`.
-    pub(crate) fn resolve_clip_sets(&self, stack: &LayerStack) -> Result<Vec<clip::ResolvedClipSet>> {
-        let mut sets: HashMap<String, HashMap<String, Value>> = HashMap::new();
-        let mut blocked_sets: HashSet<String> = HashSet::new();
-        let mut asset_layers: HashMap<String, usize> = HashMap::new();
-        let mut manifest_layers: HashMap<String, usize> = HashMap::new();
-        // Sets with explicit `assetPaths` (whose `active`/`times` are retimed
-        // as they compose) versus the offset of a template set's authoring
-        // node (whose schedule is derived later and retimed afterwards).
-        let mut explicit_sets: HashSet<String> = HashSet::new();
-        let mut template_offsets: HashMap<String, LayerOffset> = HashMap::new();
-
-        // Opinions fan out per contributing sublayer, strongest first; a value
-        // block on any layer stops every weaker opinion (spec 12.3.4).
-        for opinion in self.opinions(FieldKey::Clips.as_str(), stack, None) {
-            let Opinion {
-                layer, value, offset, ..
-            } = opinion?;
-            match value.into_owned() {
-                Value::ValueBlock => break,
-                Value::Dictionary(dict) => {
-                    for (set_name, set_value) in dict {
-                        if blocked_sets.contains(&set_name) {
-                            continue;
-                        }
-                        let Value::Dictionary(fields) = set_value else {
-                            if !sets.contains_key(&set_name) {
-                                blocked_sets.insert(set_name);
-                            }
-                            continue;
-                        };
-                        let composed = sets.entry(set_name.clone()).or_default();
-                        for (field, value) in fields {
-                            if composed.contains_key(&field) {
-                                continue;
-                            }
-                            let value = if field == clip::keys::ACTIVE || field == clip::keys::TIMES {
-                                retime_clip_stage_times(value, offset)
-                            } else {
-                                value
-                            };
-                            // Relative clip asset paths anchor on the layer that
-                            // authored them. Explicit `assetPaths` win over a
-                            // template in parse_one, so they always set the
-                            // anchor, while `templateAssetPath` only sets it when
-                            // no explicit `assetPaths` has been composed — else a
-                            // weaker template layer would mis-anchor the explicit
-                            // paths the stronger layer authored.
-                            if field == clip::keys::ASSET_PATHS {
-                                asset_layers.insert(set_name.clone(), layer);
-                                explicit_sets.insert(set_name.clone());
-                            } else if field == clip::keys::TEMPLATE_ASSET_PATH {
-                                if !explicit_sets.contains(&set_name) {
-                                    asset_layers.insert(set_name.clone(), layer);
-                                }
-                                template_offsets.insert(set_name.clone(), offset);
-                            } else if field == clip::keys::MANIFEST_ASSET_PATH {
-                                manifest_layers.insert(set_name.clone(), layer);
-                            }
-                            composed.insert(field, value);
-                        }
-                    }
-                }
-                _ => continue,
-            }
-        }
-
-        let clips = Value::Dictionary(
-            sets.into_iter()
-                .map(|(name, fields)| (name, Value::Dictionary(fields)))
-                .collect(),
-        );
-        let order = self.clip_sets_order(stack)?;
-
-        Ok(clip::ClipSet::parse_all(&clips, order.as_deref())
-            .into_iter()
-            .filter_map(|mut set| {
-                let asset_layer = asset_layers.get(&set.name).copied()?;
-                let manifest_layer = manifest_layers.get(&set.name).copied();
-                // Explicit `active`/`times` were retimed as they composed. A
-                // template schedule is derived in clip time, so retime its
-                // stage times here by the authoring node's offset.
-                if !explicit_sets.contains(&set.name) {
-                    if let Some(&offset) = template_offsets.get(&set.name) {
-                        set.retime_stage_times(offset);
-                    }
-                }
-                Some(clip::ResolvedClipSet {
-                    set,
-                    asset_layer,
-                    manifest_layer,
-                })
-            })
-            .collect())
-    }
-
-    /// Variability resolution per spec 12.2.3: weakest authored opinion wins.
-    /// Iterates strongest-to-weakest tracking the latest match, so a
-    /// [`Value::ValueBlock`] still blocks weaker opinions.
-    fn resolve_variability(&self, stack: &LayerStack, prop_suffix: Option<&str>) -> Result<Option<Value>> {
-        let field = FieldKey::Variability.as_str();
-        let mut weakest = None;
-        for opinion in self.opinions(field, stack, prop_suffix) {
-            let value = opinion?.value;
-            if matches!(value.as_ref(), Value::ValueBlock) {
-                break;
-            }
-            if matches!(value.as_ref(), Value::Variability(_)) {
-                weakest = Some(value.into_owned());
-            }
-        }
-        Ok(weakest)
-    }
-
-    /// `custom` resolution per spec 12.2.4: any-true across authored opinions.
-    /// Returns `Bool(true)` as soon as any opinion is true, `Bool(false)` if
-    /// at least one opinion was authored but none were true, and `None`
-    /// otherwise.
-    fn resolve_custom(&self, stack: &LayerStack, prop_suffix: Option<&str>) -> Result<Option<Value>> {
-        let field = FieldKey::Custom.as_str();
-        let mut saw_opinion = false;
-        for opinion in self.opinions(field, stack, prop_suffix) {
-            let value = opinion?.value;
-            if matches!(value.as_ref(), Value::ValueBlock) {
-                break;
-            }
-            saw_opinion = true;
-            if matches!(value.as_ref(), Value::Bool(true)) {
-                return Ok(Some(Value::Bool(true)));
-            }
-        }
-        Ok(saw_opinion.then_some(Value::Bool(false)))
-    }
-
-    /// Specifier resolution per spec 12.2.1.
-    ///
-    /// `over` is undefining; `def` and `class` are defining. The composed
-    /// specifier is `def` if the strongest defining opinion is `def`, or if
-    /// the strongest defining opinion not from a direct inherit is `def`.
-    /// It is `class` if the strongest defining opinion not from a direct
-    /// inherit is `class`, or if every defining opinion is `class`. It is
-    /// `over` only when every authored opinion is `over`.
-    fn resolve_specifier(&self, stack: &LayerStack, prop_suffix: Option<&str>) -> Result<Option<Value>> {
-        let field = FieldKey::Specifier.as_str();
-        let mut specs: Vec<(Specifier, ArcType)> = Vec::new();
-        for opinion in self.opinions(field, stack, prop_suffix) {
-            let Opinion { node, value, .. } = opinion?;
-            if matches!(value.as_ref(), Value::ValueBlock) {
-                break;
-            }
-            if let Value::Specifier(s) = value.into_owned() {
-                specs.push((s, node.arc));
-            }
-        }
-        if specs.is_empty() {
-            return Ok(None);
-        }
-
-        let strongest_defining = specs.iter().find(|(s, _)| *s != Specifier::Over).map(|(s, _)| *s);
-        let Some(strongest) = strongest_defining else {
-            // All authored opinions are `over`.
-            return Ok(Some(Value::Specifier(Specifier::Over)));
-        };
-
-        let strongest_non_inherit_defining = specs
-            .iter()
-            .find(|(s, arc)| *s != Specifier::Over && *arc != ArcType::Inherit)
-            .map(|(s, _)| *s);
-
-        if strongest == Specifier::Def || strongest_non_inherit_defining == Some(Specifier::Def) {
-            return Ok(Some(Value::Specifier(Specifier::Def)));
-        }
-
-        let all_defining_class = specs
-            .iter()
-            .filter(|(s, _)| *s != Specifier::Over)
-            .all(|(s, _)| *s == Specifier::Class);
-        if strongest_non_inherit_defining == Some(Specifier::Class) || all_defining_class {
-            return Ok(Some(Value::Specifier(Specifier::Class)));
-        }
-
-        Ok(Some(Value::Specifier(strongest)))
     }
 }
 
@@ -2721,18 +1531,6 @@ impl<'a> IndexBuilder<'a> {
 // Shared helpers (used by IndexBuilder and Stage)
 // ---------------------------------------------------------------------------
 
-/// Folds list-op opinions, collected strongest-to-weakest, into a single
-/// resolved list (spec 12.2.6): each weaker op is composed under the stronger
-/// ones. `compose_over` short-circuits on an explicit op, so a stronger
-/// explicit opinion replaces all weaker contributions.
-fn compose_list_ops<T: Default + Clone + PartialEq>(ops: &[sdf::ListOp<T>]) -> Vec<T> {
-    let mut result = Vec::new();
-    for op in ops.iter().rev() {
-        result = op.compose_over(&result);
-    }
-    result
-}
-
 /// Expands an internal reference target by inserting variant segments from
 /// the context path.
 ///
@@ -2876,54 +1674,6 @@ fn resolve_variant_selections_in<'a>(
     }
 
     selections
-}
-
-/// Maps time sample keys from layer time to stage time through `offset`
-/// (spec 12.3.2.1): `stage_t = offset + scale * layer_t`. Returns the samples
-/// untouched when `offset` is the identity.
-fn retime_samples(samples: sdf::TimeSampleMap, offset: LayerOffset) -> sdf::TimeSampleMap {
-    if offset.is_identity() {
-        return samples;
-    }
-    samples.into_iter().map(|(t, value)| (offset.apply(t), value)).collect()
-}
-
-/// Maps the stage-time component of clip `active`/`times` pairs through the
-/// layer offset of the node that authored the field.
-fn retime_clip_stage_times(value: Value, offset: LayerOffset) -> Value {
-    if offset.is_identity() {
-        return value;
-    }
-    match value {
-        Value::Vec2dVec(pairs) => Value::Vec2dVec(
-            pairs
-                .into_iter()
-                .map(|[stage, data]| [offset.apply(stage), data])
-                .collect(),
-        ),
-        other => other,
-    }
-}
-
-/// Applies `strong over weak` dictionary composition in place.
-///
-/// Keys authored in the stronger dictionary win. If both dictionaries hold a
-/// dictionary at the same key, those nested dictionaries are composed
-/// recursively; otherwise the stronger value is kept.
-fn dictionary_over(stronger: &mut HashMap<String, Value>, weaker: HashMap<String, Value>) {
-    for (key, weaker_value) in weaker {
-        match stronger.entry(key) {
-            Entry::Occupied(mut entry) => {
-                if let (Value::Dictionary(strong_dict), Value::Dictionary(weak_dict)) = (entry.get_mut(), weaker_value)
-                {
-                    dictionary_over(strong_dict, weak_dict);
-                }
-            }
-            Entry::Vacant(entry) => {
-                entry.insert(weaker_value);
-            }
-        }
-    }
 }
 
 /// Composes a list-op field across nodes' layer stacks, returning the
@@ -4249,21 +2999,6 @@ def "Prim" (
             "missing /Model/Anim at effective (30, 1): got {got:?}"
         );
         Ok(())
-    }
-
-    #[test]
-    fn retime_samples_offset_scale() {
-        let samples: sdf::TimeSampleMap = vec![(1.0, Value::Double(0.0)), (5.0, Value::Double(1.0))];
-        let retimed = retime_samples(samples, LayerOffset::new(10.0, 2.0));
-        let times: Vec<f64> = retimed.iter().map(|(t, _)| *t).collect();
-        assert_eq!(times, vec![12.0, 20.0]);
-    }
-
-    #[test]
-    fn retime_samples_identity_passthrough() {
-        let samples: sdf::TimeSampleMap = vec![(1.0, Value::Double(0.0))];
-        let retimed = retime_samples(samples.clone(), LayerOffset::IDENTITY);
-        assert_eq!(retimed, samples);
     }
 
     /// `resolve_time_samples` must map authored layer times to stage times
